Seurat: check if mito prefix is valid
The filtering of cells might fail if no cells are found that match the
filtering criteria. One way this can happen, is when an invalid
mitochondrial-gene prefix is provided while also filtering on the
percent mitochondrial genes.
By adding a check for the prefix, we can provide the user with a more
meaninful error instead of the default Seurat error.